+++ contacts/app/controllers/favorites_controller.rb
@@ -1,10 +1,14 @@
 class FavoritesController < ApplicationController
+
+  before_filter :authenticate_user
+
   def index
-    fav_contacts = User.find(params[:user_id]).favorite_contacts
+    fav_contacts = current_user.favorite_contacts
     render :json => { 'favorites' => fav_contacts}
   end
 
   def create
+    params[:favorite][:user_id] = current_user.id
     favorite = Favorite.new(params[:favorite])
     if favorite.save
       render :json => favorite
@@ -15,7 +19,11 @@ class FavoritesController < ApplicationController
 
   def destroy
     favorite = Favorite.find(params[:id])
-    favorite.destroy
-    render :json => "Success. Favorite deleted."
+    if favorite.user == current_user
+      favorite.destroy
+      render :json => "Success. Favorite deleted."
+    else
+      render :json => "Can't delete others' favorites", :status => :forbidden
+    end
   end
 end
